+++ src/DiceBotLoaderList.js
@@ -1,10 +1,11 @@
-import { DiceBotLoaderList as RubyDiceBotLoaderList } from '../dist/bcdice.ruby.js';
+import '../lib/bcdice.ruby.js';
+import opal from './opal';
 import DiceBotLoader from './DiceBotLoader';
 import { isNil } from './utilities';
 
 export default class DiceBotLoaderList {
     static find(gameTitle) {
-        const loader = RubyDiceBotLoaderList.$find(gameTitle);
+        const loader = opal(Opal => Opal.DiceBotLoaderList.$find(gameTitle));
         if (isNil(loader)) return null;
 
         return new DiceBotLoader(loader);