+++ tests/DiceBotLoaderList.test.js
@@ -2,7 +2,7 @@ describe('DiceBotLoaderList', () => {
     var DiceBotLoaderList;
 
     it('is valid module', () => {
-        DiceBotLoaderList = require('../dist/DiceBotLoaderList').default;
+        DiceBotLoaderList = require('../lib/DiceBotLoaderList').default;
     });
 
     it('is defined', () => {
@@ -11,12 +11,12 @@ describe('DiceBotLoaderList', () => {
 
     describe('.find', () => {
         it('returns loader', () => {
-            const DiceBotLoader = require('../dist/DiceBotLoader').default;
+            const DiceBotLoader = require('../lib/DiceBotLoader').default;
 
             const loader = DiceBotLoaderList.find('COC');
             expect(loader).toBeInstanceOf(DiceBotLoader);
 
-            const DiceBot = require('../dist/DiceBot').default;
+            const DiceBot = require('../lib/DiceBot').default;
             const diceBot = loader.loadDiceBot();
             expect(diceBot).toBeInstanceOf(DiceBot);
             expect(diceBot.gameType()).toEqual('Cthulhu');