+++ tests/index.test.js
@@ -6,22 +6,22 @@ describe('index.js', () => {
     });
 
     it('exports BCDice as default', () => {
-        expect(Index.default).toBe(require('../dist/BCDice').default);
+        expect(Index.default).toBe(require('../lib/BCDice').default);
     });
 
     it('exports BCDice', () => {
-        expect(Index.BCDice).toBe(require('../dist/BCDice').default);
+        expect(Index.BCDice).toBe(require('../lib/BCDice').default);
     });
 
     it('exports DiceBot', () => {
-        expect(Index.DiceBot).toBe(require('../dist/DiceBot').default);
+        expect(Index.DiceBot).toBe(require('../lib/DiceBot').default);
     });
 
     it('exports DiceBotLoader', () => {
-        expect(Index.DiceBotLoader).toBe(require('../dist/DiceBotLoader').default);
+        expect(Index.DiceBotLoader).toBe(require('../lib/DiceBotLoader').default);
     });
     
     it('exports DiceBotLoaderList', () => {
-        expect(Index.DiceBotLoaderList).toBe(require('../dist/DiceBotLoaderList').default);
+        expect(Index.DiceBotLoaderList).toBe(require('../lib/DiceBotLoaderList').default);
     });
 });