+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Fix destruction of Function.prototype
+### Changed
+- Rename `dist` to `lib`
+
 ## [0.2.0] - 2017-08-22
 ### Added
 - Added wrapper classes around Loaders